➕ Dependency: Change Dependency From MySQL Connector To MariaDB [main]
Changes Made:
	- Changed Database Connector from MySQL Connector to MariaDB Connector.

MySQL Connector does not work on my current system due to compatibility
issues. So I switched to MariaDB Connector which is a connector for
MariaDB a drop-in replacement for MySQL. It works perfectly fine with
MySQL databases as well. All the code changes were related to changing
the import statements and connection method.

Files Changed:
	modified:   MediCheck.py
	modified:   poetry.lock
	modified:   pyproject.toml

Date: 17/11/2025, 02:33 P.M. IST

All Done.

Signed-off-by: Kunal Kumar <kunalkumarchourasiya2021@gmail.com>